Corrected a single line incompatible with Godot 4.3+
Upon using the shader on Godot 4.3+ (tested on mono version), an error occurs:

Line 48 (MAGIC_POSITION_WRITE):You are attempting to assign the VERTEX position in model space to the vertex POSITION in clip space. The definition of clip space changed in version 4.3, so if this code was written prior to 4.3, it will not continue to work. Consider specifying the clip space z-component directly i.e. use `vec4(VERTEX.xy, 1.0, 1.0)`.

That recommended change works flawlessly.